% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPBT.R
\name{CPBT}
\alias{CPBT}
\title{Coastal Protection Benefit Tool}
\format{
Patch attributes must include.
\describe{
  \item{hc}{Numeric, Blade height in meters}
  \item{N}{Numeric, Shoot density as number of shoots per meter squared.}
  \item{d}{Numeric, Blade width in meters.}
  \item{Type}{Either 'Eelgrass', 'Kelp' or 'Marsh'.}
  \item{Cd}{Numeric, Drag coefficent as per Guannel et al 2015.
  If unsure use a value of 0.1 for Eelgrass}
}

foreshore beach attributes. Each beach section should be characterized with field attributes:
\describe{
  \item{slope}{the foreshore slope (slope) as rise over run.}
  \item{W}{Numeric, the berm width in meters (W).}
  \item{B}{Numeric, berm height in meters (B).}
  \item{D}{Numeric, dune height (D) in meters.}
  \item{sediment}{Numeric, the sediment grain size in
millimeters (sediment).}
  \item{V}{Numeric, the beach value per meter squared in dollars (V).}
}

Building footprint attribute fields must include columns:
\describe{
  \item{DDID}{Numeric, the HAZUS depth-damage curves IDs populated for
each structure (column name: DDID).}
  \item{VAL}{Numeric, estimates of the replacement value cost
for each structure in dollars (VAL).}
}
}
\usage{
CPBT(
  simulation_name = "My Simulation",
  Scenario_Description = NA,
  dir_output = getwd(),
  Coastline = NA,
  ShorelinePointDist = 200,
  BufferDist = 50,
  RadLineDist = 1,
  TopoBathy = NA,
  SmoothParameter = 5,
  MaxOnshoreDist = 0.1,
  trimline = NA,
  Vegetation = NA,
  mangrove = NULL,
  mean_high_water = 1.5,
  mean_sea_level = 0.5,
  tide_during_storm = 0.9,
  surge_elevation = 0.5,
  sea_level_rise = 0.15,
  Ho = 2.5,
  To = 7,
  storm_duration = 3,
  BeachAttributes = NA,
  Tr = 10,
  PropValue = 200,
  disc = 0.05,
  TimeHoriz = 100,
  Bldgs = NA,
  export_report = FALSE,
  export_spatial_dat = FALSE
)
}
\arguments{
\item{simulation_name}{Simulation name for current storm and conditions.}

\item{Scenario_Description}{A qualitative description of the scenario.}

\item{dir_output}{Local output directory for CPBT simulation results.}

\item{Coastline}{Linestring of class sf and dataframe. Coastline spatial
line segment for modeling. The coastline should roughly trace the water
edge along the beach at the mean sea level. It is recommended to keep the coastline geometry simple and avoid sharp corners.}

\item{ShorelinePointDist}{Numeric. Spacing between cross-shore profile lines (meters).
Note that it is recommended to keep this number as large as possible for
your area of interest to speed up processing time.}

\item{BufferDist}{Numeric. Buffer search distance (meters) used to identify
the perpendicular angle of each cross-shore profile line from the coastline
(recommended values are 100-300).}

\item{RadLineDist}{Numeric. Radial line distance (kilometers) of the cross-shore profiles. This value determines how for offshore (and onshore) the cross-shore profiles should extend (recommended values are 1 - 3 km).}

\item{TopoBathy}{TopoBathy digital elevation model of class RasterLayer.}

\item{SmoothParameter}{Numeric, smoothing window length as a percentage of
the cross-shore profile length (0-100). A value of zero means no smoothing, but see
details below (recommended values are 5-20).}

\item{MaxOnshoreDist}{Numeric. Maximum radial line distance onshore in kilometers.
Note that in some instances the onshore extent should be truncated
significantly for the wave evolution model. It is recommended to keep this
value below 1km.}

\item{trimline}{Optional back shore trim line used to restrict the on-shore
extent of the cross-shore profiles. The default value is NA (not used). If
used, the back shore trim line should run parallel to the coastline but be
set back onto land. A back shore trim line may be required in cases where
cross-shore profiles are generated along the coastline of a narrow peninsula
or in instances where there is a back shore lagoon. A back shore trimline
should be provided as a simple features line object.}

\item{Vegetation}{A simple feature polygon of class sf and dataframe.}

\item{mangrove}{Named numeric vector. Defaults to NULL. See details in `WaveModel`}

\item{mean_high_water}{Mean high tidal water elevation above the chart datum.}

\item{mean_sea_level}{Mean average tidal water elevation above the chart datum.}

\item{tide_during_storm}{Numeric, Tidal elevation in meters during the storm referenced to the Chart Datum.}

\item{surge_elevation}{Numeric, Additional elevation from storm surge (in meters) during the storm.}

\item{sea_level_rise}{Numeric, Additional elevation from any local sea-level rise (in meters) during the storm.}

\item{Ho}{Initial offshore wave height in meters.}

\item{To}{Initial offshore wave period in seconds.}

\item{storm_duration}{Numeric. Storm duration in hours.}

\item{BeachAttributes}{Spatial polygons of class sf and dataframe for}

\item{Tr}{Numeric. Return period (frequency) of the simulated storm (in years).}

\item{PropValue}{Generaly land value in dollars per square meter of beach.
(not used if beach polygon property values are provided).}

\item{disc}{Annual valuation discount rate over the time horizon (0 - 1).}

\item{TimeHoriz}{Time horizon (in years) for long term cumulative valuation given a storm return frequency.
Typically 100-year horizons are used.}

\item{Bldgs}{Spatial polygon of building footprints of class sf and dataframe.
See data(Bldgs) for an example input format.}

\item{export_report}{Boolean for debugging. Keep set at TRUE.}

\item{export_spatial_dat}{Boolean (TRUE FALSE) should spatial data be exported for report.}
}
\value{
Export a html output report folder to a local directory with the
complete CPBT results summary.
}
\description{
Runs the complete CPBT analysis along a coastline for a given
storm simulation.
}
\examples{
\dontrun{

library(MNAI.CPBT)

data(Coastline)
data(BeachAttributes)
rpath <- system.file("extdata", "TopoBathy.tif", package = "MNAI.CPBT")
TopoBathy <- raster::raster(rpath)
data(Vegetation)
data(Bldgs)

CPBT(
  simulation_name = "My Simulation",
  dir_output = getwd(),
  Coastline = Coastline,
  ShorelinePointDist = 50,
  BufferDist = 25,
  RadLineDist = 1,
  TopoBathy = TopoBathy,
  SmoothParameter = 5,
  MaxOnshoreDist = 0.1,
  trimline = NA,
  Vegetation = Vegetation,
  mean_high_water = 1.5,
  mean_sea_level = 0.5,
  tide_during_storm = 0.9,
  surge_elevation = 0.2,
  sea_level_rise = 0.15,
  Ho = 3,
  To = 9,
  storm_duration = 6,
  BeachAttributes = BeachAttributes,
  Tr = 50,
  PropValue = 200,
  disc = 0.03,
  TimeHoriz = 200,
  Bldgs = Bldgs,
  export_report = FALSE
)

}
}
