% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErosionTransect.R
\name{ErosionTransect}
\alias{ErosionTransect}
\title{Custom ErosionTransect}
\usage{
ErosionTransect(
  wave_transect = NA,
  Ho = NA,
  To = NA,
  sediment_size_mm = NA,
  ssf_lookup = NA,
  total_wsl_adj = NA,
  m = 0.02,
  B = 3,
  D = 3,
  W = 20,
  storm_duration = NA
)
}
\description{
Custom ErosionTransect
}
\keyword{internal}
