% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErosionTransectMangrove.R
\name{ErosionTransectMangrove}
\alias{ErosionTransectMangrove}
\title{Custom ErosionTransectMangrove}
\usage{
ErosionTransectMangrove(
  wave_transect = wave_transect,
  Ho = Ho,
  To = To,
  total_wsl_adj = total_wsl_adj,
  m = m,
  me = me,
  Cm = Cm,
  storm_duration = storm_duration
)
}
\description{
Custom ErosionTransectMangrove
}
\keyword{internal}
