% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportJSONContent.R
\name{ExportJSONContent}
\alias{ExportJSONContent}
\title{Exports JSON content}
\usage{
ExportJSONContent(
  path_output = NA,
  flood_contour = NA,
  merg_Veg = NA,
  wave_dat = NA,
  dat = NA,
  ero_tot = NA,
  total_wsl_adj = NA
)
}
\description{
Exports JSON content
}
\keyword{internal}
