% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportProfiles.R
\name{ExportProfiles}
\alias{ExportProfiles}
\title{ExportProfiles}
\usage{
ExportProfiles(
  dat = NA,
  wave_dat = NA,
  erosion = NA,
  mean_high_water = NA,
  total_wsl_adj = NA,
  TrimOnshoreDist = NA,
  path_output = NA,
  export_csv = TRUE
)
}
\arguments{
\item{dat}{A sf and data frame of cross-profiles returned from ExtractVeg.}

\item{wave_dat}{A sf and data frame of cross-profiles returned from
WaveModel.}

\item{erosion}{Erosion summaries returned from ErosionTransectsUtil.}

\item{mean_high_water}{Mean high water elevation above the chart datum.}

\item{total_wsl_adj}{Total water surface level above the chart datum. Recall
that the chart datum and TopoBathy DEM are referenced to have 0 at low
water. It is therefore suggested to set this value at the mean sea level
above chart datum or a specific tidal elevation of interest.}

\item{TrimOnshoreDist}{Numeric, option distance onshore to trim profile in
meters.}

\item{path_output}{Output directory to save cross-shore profiles.}

\item{export_csv}{Should cross shore profiles be exported.}
}
\value{
Cross-shore profiles are exported and save as csv to a directory.
}
\description{
Exports cross shore profiles as csv files to a specified local
directory.
}
