% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FloodVulnerabilityStudy.R
\name{FloodVulnerabilityStudy}
\alias{FloodVulnerabilityStudy}
\title{Flood Vulnerability Study}
\usage{
FloodVulnerabilityStudy(
  Bldgs = NA,
  group_col = NA,
  TopoBathy = NA,
  start_elevation = 2,
  end_elevation = 10,
  step_size = 0.2,
  simulation_name = "My Simulation",
  dir_output = getwd()
)
}
\arguments{
\item{Bldgs}{Spatial polygon of building footprints of class sf and dataframe.
See data(Bldgs) for an example input format.}

\item{group_col}{(Optional) Column name of group column name in building
data frame. If used results will be summarized by region.}

\item{TopoBathy}{TopoBathy digital elevation model of class RasterLayer.}

\item{start_elevation}{Starting lower water level elevation for sequence.}

\item{end_elevation}{Ending upper level elevation for sequence.}

\item{step_size}{Water level elevation step size for sequence.}

\item{simulation_name}{Character. Simulation name for export.}

\item{dir_output}{Local output directory for CPBT simulation results.}
}
\value{
Exports results to a specified folder file path.
}
\description{
Calculates the number of flooded structures, median flood
depth and total structural damage across a range of water level elevations.
}
\details{
The flood vulnerability is a computationally intensive yet simple
tool to allow communities to better understand the vulnerability of their
communities to coastal flooding. The function works by gradually raising
the water level by a fixed amount and then recalculating the structural
damage cost of a flood at that water level. Users define a range of
elevation e.g., 0 – 5 meters above chart datum and a step size e.g.,
0.15 m and then the function calculates to total damage cost at each
0.15 interval. A html report is exported with three plots showing the
depth-damage curves used in the assessment, the number of affected
structures at each water level and the total damage cost at each
water level. The intent of this tool is to look at overall damage cost
to different storms. If the number of affected structures and total
damage costs quickly climb upwards at low water levels, then users can
assume that the area is highly vulnerable to coastal flooding.
Conversely, if damage costs do not show any significantly values
until the water level reaches extreme levels, then users can
conclude that the community is naturally resilient to flooding.
Users can also evaluate tipping points at which the damage cost
climbs significantly for a given water level. For example, if a
community is expect to have no structural damage below flood water
levels of 2.1 m CD, but then experience millions of dollars in
damages between water levels of 2.1 – 2.5m then users can conclude
that any forces pushing the water levels above 2 m represent a
significant concern for their communities.
}
