% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterativek.R
\name{iterativek}
\alias{iterativek}
\title{Calculates the wave number}
\usage{
iterativek(sigma, dh)
}
\arguments{
\item{sigma}{a number, angular frequency}

\item{dh}{a number, water depth}
}
\description{
Calculates the wave number
}
\keyword{internal}
