% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkProfilesToBeaches.R
\name{LinkProfilesToBeaches}
\alias{LinkProfilesToBeaches}
\title{Link Profiles To Beaches}
\usage{
LinkProfilesToBeaches(BeachAttributes = NA, dat = NA)
}
\arguments{
\item{BeachAttributes}{class sf and dataframe of foreshore beach attributes.}

\item{dat}{Cross-shore profile spatial points object returned from WaveModel.}
}
\value{
A data table of foreshore attributes for each cross shore profile.
}
\description{
Merges the profiles returned from WaveModel and adds foreshore
beach attributes to them.
}
