% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MudErosion.R
\name{MudErosion}
\alias{MudErosion}
\title{Mud Erosion}
\usage{
MudErosion(Uc = NA, Uw = NA, h = NA, To = NA, me = 0.001, Cm = 70)
}
\arguments{
\item{Uc}{Vector of zeros}

\item{Uw}{Bottom velocity vector (m/s)}

\item{h}{Depth vector (m)}

\item{To}{Wave period in seconds}

\item{me}{Erosion constant. Default value is 0.001m/s (kg/Ns)}

\item{Cm}{Dry density of bed. Default value is 70kg/m3}
}
\description{
Calculates erosion for mudflat transect.
}
\keyword{internal}
