% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePoints.R
\name{samplePoints}
\alias{samplePoints}
\title{samplePoints}
\usage{
samplePoints(
  Coastline = NA,
  ShorelinePointDist = 200,
  BufferDist = 300,
  RadLineDist = 3
)
}
\arguments{
\item{Coastline}{Coastline LINESTRING of class sf and data.frame.}

\item{ShorelinePointDist}{Spacing between cross-shore profile lines (meters).
Note that it is recommended to keep this number as large as possible for
your area of interest to speed up processing time.}

\item{BufferDist}{Buffer search distance (meters) used to identify
the perpendicular angle of each cross-shore profile line from the coastline.}

\item{RadLineDist}{Radial line distance (meters) of the cross-shore profiles.}
}
\value{
A list object of shoreline points and cross-shore profile lines
along the coastline.
}
\description{
Sample points along the coastline at specified intervals.
}
\examples{
\dontrun{
require(sf)
data(Coastline)
shoreline_points <- samplePoints(
  Coastline = Coastline,
  ShorelinePointDist = 150, # where to sample points on shoreline
  BufferDist = 300.0, # Buffer distance in meters for vertical line
  RadLineDist = 3.0 # Line dist in km
)
# Coastline
plot(sf::st_geometry(Coastline))
# Shoreline points
plot(sf::st_geometry(shoreline_points[[1]]), add=TRUE)
# Cross-shore profiles
plot(sf::st_geometry(shoreline_points[[2]]), add=TRUE)
}
}
