% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignalSmooth_smooth.R
\name{SignalSmooth_smooth}
\alias{SignalSmooth_smooth}
\title{Internal SignalSmooth_smooth}
\usage{
SignalSmooth_smooth(x, window_len, mm = "hanning")
}
\arguments{
\item{x}{a cross shore profile elevaion vector}

\item{window_len}{length of the smoothing window}

\item{mm}{method. currently only hanning available.}
}
\description{
Internal SignalSmooth_smooth
}
\keyword{internal}
