% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submergedStructure.R
\name{submergedStructure}
\alias{submergedStructure}
\title{Submerged Structure Analysis}
\usage{
submergedStructure(Hi = 2.5, To = 7, hi = 6, hc = 3, Cwidth = 5, Bwidth = 10)
}
\arguments{
\item{Hi}{Initial wave height (m) at the seaward margin of a submerged
structure.}

\item{To}{Initial wave period (s) at the seaward margin of a submerged
structure.}

\item{hi}{Total water depth (m) at the seaward margin of a submerged
structure.}

\item{hc}{Height of the submerged structure (m) above the surrounding seabed.}

\item{Cwidth}{Crest width (m) at the top of the submerged structure.}

\item{Bwidth}{Base width (m) at the bottom of the submerged structure.}
}
\value{
The final wave height after passing over a submerged structure
is reported as a fraction (0 to 1) or the original wave height.
}
\description{
Estimates the reduction in wave height as a wave passes over a
submerged structure.
}
\examples{
pfinal <- submergedStructure(Hi = 2.5, To = 7, hi = 6, hc = 3, Cwidth = 2,
Bwidth = 5)
2.5*pfinal       # Wave height after passing over submerged structure.
2.5-(2.5*pfinal) # Wave height reduction in meters.
(1-pfinal)*100   # Wave height reduction percentage.
}
