% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoBathy-data.R
\docType{data}
\name{TopoBathy}
\alias{TopoBathy}
\title{Example TopoBathy Raster Layer}
\format{
An object of class \code{"RasterLayer"};.
}
\source{
\href{https://www.open.canada.ca/en}{CHS Bathymetry Data Download}
}
\usage{
data(TopoBathy)
}
\description{
Sample TopoBathy (topographic and bathymetric) raster layer
used for function examples. 0 meters is referenced to the
Mean Lower Low Water (MLLW) level (Chart Datum).
Note that this dataset has been reduced and simplified from
the original version and should only be used for demonstration purposes.
See the source link below to download relevant data for your region of
interest.
}
\examples{
\donttest{
require(raster)
data(TopoBathy)
raster::plot(TopoBathy)
}
}
\references{
DFO - Canadian Hydrographic Service Non-Navigational (NONNA)
Bathymetric Data
}
\keyword{datasets}
