% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimline-data.R
\docType{data}
\name{Trimline}
\alias{Trimline}
\title{Trimline}
\format{
A simple feature LINESTRING of class sf and data.frame.
}
\usage{
data(Trimline)
}
\description{
The back shore trim line is optionally included to restrict the on-shore
extent of the cross-shore profiles. The trim line runs parallel to the
coastline but is drawn further upslope on the land away from the water. A
back shore trim line may be required in cases where cross-shore profiles
are generated along the coastline of a narrow peninsula or in instances
where there is a back shore lagoon. A back shore trimline should be
provided as a simple features line object.
}
\examples{
\donttest{
require(sf)
data(Trimline)
# Plot the trimline
plot(sf::st_geometry(Trimline))
}
}
\keyword{datasets}
