% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveModel.R
\name{WaveModel}
\alias{WaveModel}
\title{WaveModel}
\usage{
WaveModel(
  dat = NA,
  total_wsl_adj = NA,
  Ho = NA,
  To = NA,
  tran_force = FALSE,
  print_debug = FALSE
)
}
\arguments{
\item{dat}{A sf and dataframe of cross-profiles returned from ExtractVeg.}

\item{total_wsl_adj}{Total water surface level above the chart datum. Recall
that the chart datum and TopoBathy DEM are referenced to have 0 at low
water. It is therefore suggested to set this value at the mean sea level
above chart datum or a specific tidal elevation of interest.}

\item{Ho}{Initial offshore wave height in meters.}

\item{To}{Initial offshore wave period in second.}

\item{tran_force}{TRUE/FALSE should transect be forced even if there are
error codes.}

\item{print_debug}{Turn on function debugging mode.}
}
\value{
An object of class sf and data.frame updated with wave data
along each cross-shore profile.
}
\description{
Wave attenuation model originally developed by Dr. Greg Guannel
for the Coastal Natural Capital InVEST project. This function models wave
attenuation along a cross-shore elevation profile.
}
\details{
Wave attenuation model originally developed by Dr. Greg Guannel
for the Coastal Natural Capital InVEST project. This function models wave
attenuation along a cross-shore elevation profile. Input parameters include
a pre-processed cross shore profile dataset (dat) developed through the
sequence of functions show in the example, the total water surface level
of still water (total_wsl_adj) above the chart datum, the wave height (Ho)
and wave period (To). After running the function populates the original
input dataset (dat) with the follow data attributes: wave setup with
vegetation (Eta), wave setup without vegetation (Etas), bottom orbital
velocity (Ubot), wave height with vegetation (H_veg), wave height without
vegetation (H_noveg) and wave dissipation (Dis1). If  you receive a message
stating Transect failed - bad sort order, we suggest decreasing the
MaxOnshoreDist, adding a trim line or decreasing the water level.
}
\examples{
\dontrun{
library(MNAI.CPBT)
data(Coastline)
# Generate cross-shore profile lines along the coastline.
crossshore_profiles <- samplePoints(
  Coastline = Coastline,
  ShorelinePointDist = 150,
  BufferDist = 50,
  RadLineDist = 1.5
)
crossshore_lines <- crossshore_profiles[[2]]

# Extract elevation values along each profile
rpath <-  system.file("extdata", "TopoBathy.tif", package = "MNAI.CPBT")
TopoBathy <- raster::raster(rpath)
pt_elevs <- ExtractElev(crossshore_lines, TopoBathy)

# Run SignalSmooth function to smooth elevation profiles
pt_elevs <- SignalSmooth(point_elev = pt_elevs,
SmoothParameter = 5)

# Clean the cross-shore profiles with CleanTransect
data(Trimline)
cleantransect <- CleanTransect(
  point_elev = pt_elevs,
  RadLineDist = 1.5,
  MaxOnshoreDist = 0.01,
  trimline = Trimline
)

# Merge vegetation onto lines
data(Vegetation)
dat_veg <- ExtractVeg(pt_exp = cleantransect, Vegetation = Vegetation)

# Run the wave evolution model
wave_data <- WaveModel(dat = dat_veg,
  total_wsl_adj = 1.5,
  Ho = 2,
  To = 8
)

# Preview individual transect
dsub <- wave_data[wave_data$line_id == 2, ]

# Plot cross-shore profile
par(mfrow = c(2, 1))
plot(dsub$Xpos, dsub$elev, type = "l", xlab = "Cross-shore Distance (m)",
ylab = "Elevation (m) Chart Datum", main = "ELEVATION PROFILE")
points(dsub$Xpos, dsub$elev_smooth, col="red", type = "l")

# Add MLLW water line
abline(h = 0, lty = 2, col = "blue")

# Look at the wave height (without vegetation)
plot(dsub$Xpos, dsub$H_noveg, type = 'l', xlab = "Cross-shore Distance (m)",
ylab = "Wave Height (m)", main = "WAVE ATTENUATION")
# Add on wave height with vegetation
points(dsub$Xpos, dsub$H_veg, col="green", type = "l")

}
}
\references{
InVEST: Wave Attenuation & Erosion Reduction: Coastal Protection (G. Guannel)

Guannel et al. (2014) Integrated modeling framework to quantify the coastal
protection services supplied by vegetation. Journal of
Geophysical Research: Oceans. DOI: 10.1002/2014JC009821.
}
