% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveModelSimple.R
\name{WaveModelSimple}
\alias{WaveModelSimple}
\title{Custom WaveModelSimple from InVEST tool}
\usage{
WaveModelSimple(X, h, Ho, To, Cf = 0.01)
}
\arguments{
\item{X}{a number, distance}

\item{h}{a number, depth}

\item{Ho}{a number, wave height}

\item{To}{a number, wave period}

\item{Cf}{a number, coef}
}
\description{
Custom WaveModelSimple from InVEST tool
}
\keyword{internal}
