% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windWave.R
\name{windWave}
\alias{windWave}
\title{windWave}
\usage{
windWave(WindSpeed = 20, FetchDistance = 5000, WaterDepth = 100)
}
\arguments{
\item{WindSpeed}{Wind speed in units of meters per second.}

\item{FetchDistance}{Fetch distance in units of meters.}

\item{WaterDepth}{Average offshore water depth in units of meters.}
}
\value{
An estimate of the offshore wave height (meters) and wave period
(seconds).
}
\description{
Estimates the wave height and period for a given wind speed,
fetch distance and water depth.
}
\examples{
waveHeightPeriod <- windWave(WindSpeed = 20, FetchDistance = 5000,
WaterDepth = 50)
print(waveHeightPeriod)
}
