% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErosionTotals.R
\name{ErosionTotals}
\alias{ErosionTotals}
\title{Erosion Totals}
\usage{
ErosionTotals(wave_data = NA, erosion = NA, Longshore = NA)
}
\arguments{
\item{wave_data}{A sf and dataframe of cross-profiles returned from WaveModel.}

\item{erosion}{Erosion summary table returned from ErosionTransectsUtil.}

\item{Longshore}{Longshore distance in meters should match
ShorelinePointDist used in samplePoints.}
}
\value{
A list object of erosion summaries: erosion_points and totals.
}
\description{
Summarizes erosion totals and damage summaries across the
coastline section.
}
