% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractElev.R
\name{ExtractElev}
\alias{ExtractElev}
\title{ExtractElev}
\usage{
ExtractElev(cross_shore_profiles = NA, TopoBathy = NA)
}
\arguments{
\item{cross_shore_profiles}{Cross-shore profile lines from samplePoints.}

\item{TopoBathy}{TopoBathy digital elevation model of class RasterLayer.}
}
\value{
Spatial points of class sf with extracted elevation values along
each cross-shore profile.
}
\description{
Extract elevation values along cross-shore profiles lines.
}
\examples{
\dontrun{
library(MNAI.CPBT)
data(Coastline)
shoreline_points <- samplePoints(
  Coastline = Coastline,
  ShorelinePointDist = 150, # where to sample points on shoreline
  BufferDist = 300.0, # Buffer distance in meters for vertical line
  RadLineDist = 3.0 # Line dist in km
)

# Cross-shore profile lines
plot(sf::st_geometry(Coastline))
cross_shore_profiles <- shoreline_points[[2]]
plot(sf::st_geometry(cross_shore_profiles), add=TRUE)

# Extract elevation along each profile
rpath <-  system.file("extdata", "TopoBathy.tif", package = "MNAI.CPBT")
TopoBathy <- raster::raster(rpath)
pt_elevs <- ExtractElev(cross_shore_profiles, 1, TopoBathy)
plot(pt_elevs['elev'])
# plot(TopoBathy)
}
}
